requirements.txt
django>=1.7
pyyaml>=4.2b4
dicttoxml>=1.7.4
